+++ core/auth/auth.go
@@ -553,9 +553,17 @@ func (a *Auth) createLoginSession(anonymous bool, sub string, authType model.Aut
 		return nil, errors.WrapErrorAction(logutils.ActionCreate, logutils.TypeToken, nil, err)
 	}
 
+	now := time.Now().UTC()
+	var forceExpires *time.Time
+	if appType.Application.MaxLoginSessionDuration != nil {
+		forceLogoutTime := now.Add(time.Duration(*appType.Application.MaxLoginSessionDuration) * time.Hour)
+		forceExpires = &forceLogoutTime
+	}
+
 	loginSession := model.LoginSession{ID: id, AppOrg: appOrg, AuthType: authType,
 		AppType: appType, Anonymous: anonymous, Identifier: sub, AccountAuthType: accountAuthType,
-		Device: device, IPAddress: ipAddress, AccessToken: accessToken, RefreshTokens: []string{refreshToken}, Params: params, Expires: *expires, DateCreated: time.Now()}
+		Device: device, IPAddress: ipAddress, AccessToken: accessToken, RefreshTokens: []string{refreshToken}, Params: params,
+		Expires: *expires, ForceExpires: forceExpires, DateCreated: now}
 
 	return &loginSession, nil
 }